Code clean up: * Prevents concurrent modification exception during server shutdown if a transaction is active. * Prevents NPE is internal tries to remove a cached parsed data model but has already been removed. * Prevents NPE if a collection contains a null value when reading a list of FK (not expected). * Javadoc fix.
git-svn-id: http://talendforge.org/svn/tom/trunk@106503 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.mdm.core/src/com/amalto/core/server/MetadataRepositoryAdminImpl.java
@@ -161,11 +161,13 @@ class MetadataRepositoryAdminImpl implements MetadataRepositoryAdmin {
         assertMetadataRepositoryId(metadataRepositoryId);
         synchronized (metadataRepository) {
             MetadataRepository repository = get(metadataRepositoryId);
-            repository.close();
-            metadataRepository.remove(metadataRepositoryId);
-            if (!metadataRepositoryId.endsWith(StorageAdmin.STAGING_SUFFIX)) {
-                // Remove staging metadata repository.
-                remove(metadataRepositoryId + StorageAdmin.STAGING_SUFFIX);
+            if (repository != null) {
+                repository.close();
+                metadataRepository.remove(metadataRepositoryId);
+                if (!metadataRepositoryId.endsWith(StorageAdmin.STAGING_SUFFIX)) {
+                    // Remove staging metadata repository.
+                    remove(metadataRepositoryId + StorageAdmin.STAGING_SUFFIX);
+                }
             }
         }
     }
